chore(admin): convert useMenu
