+++ packages/core/admin/admin/src/hooks/index.js
@@ -1,5 +1,4 @@
 export { useConfiguration } from './useConfiguration';
 export { useContentTypes } from './useContentTypes';
-export { default as useMenu } from './useMenu';
 export { default as useSettingsForm } from './useSettingsForm';
 export { default as useSettingsMenu } from './useSettingsMenu';
