--- packages/core/admin/admin/src/hooks/useMenu/utils/getGeneralLinks.js
@@ -1,31 +1,0 @@
-import cloneDeep from 'lodash/cloneDeep';
-
-import checkPermissions from './checkPermissions';
-
-const getGeneralLinks = async (userPermissions, generalSectionRawLinks, shouldUpdateStrapi) => {
-  const generalSectionPermissionsPromises = checkPermissions(
-    userPermissions,
-    generalSectionRawLinks
-  );
-  const generalSectionLinksPermissions = await Promise.all(generalSectionPermissionsPromises);
-
-  const authorizedGeneralSectionLinks = generalSectionRawLinks.filter(
-    (_, index) => generalSectionLinksPermissions[index]
-  );
-
-  const settingsLinkIndex = authorizedGeneralSectionLinks.findIndex(
-    (obj) => obj.to === '/settings'
-  );
-
-  if (settingsLinkIndex === -1) {
-    return [];
-  }
-
-  const authorizedGeneralLinksClone = cloneDeep(authorizedGeneralSectionLinks);
-
-  authorizedGeneralLinksClone[settingsLinkIndex].notificationsCount = shouldUpdateStrapi ? 1 : 0;
-
-  return authorizedGeneralLinksClone;
-};
-
-export default getGeneralLinks;
