+++ packages/core/admin/admin/src/pages/Admin/index.js
@@ -14,7 +14,8 @@ import { Route, Switch } from 'react-router-dom';
 
 import GuidedTourModal from '../../components/GuidedTour/Modal';
 import LeftMenu from '../../components/LeftMenu';
-import { useConfiguration, useMenu } from '../../hooks';
+import { useConfiguration } from '../../hooks';
+import { useMenu } from '../../hooks/useMenu';
 import AppLayout from '../../layouts/AppLayout';
 import { createRoute } from '../../utils';
 import { SET_APP_RUNTIME_STATUS } from '../App/constants';
