+++ packages/core/helper-plugin/src/features/RBAC.tsx
@@ -10,11 +10,16 @@ type Permission = domain.permission.Permission;
  * -----------------------------------------------------------------------------------------------*/
 
 type RBACContextValue = {
-  allPermissions?: Permission[]; // The permissions of the current user.
-  refetchPermissions?: QueryObserverBaseResult<Permission[]>;
+  allPermissions: Permission[]; // The permissions of the current user.
+  refetchPermissions: QueryObserverBaseResult<Permission[]>['refetch'];
 };
 
-const RBACContext: React.Context<RBACContextValue> = React.createContext({});
+const RBACContext = React.createContext<RBACContextValue>({
+  allPermissions: [],
+  refetchPermissions: async () => {
+    throw new Error('RBACContext: refetchPermissions() not implemented');
+  },
+});
 
 /**
  * @deprecated Use RBACContext instead.
